predicts pulling
